*稍微修改了一下DX11下Material的实现，因为有其他工作，这部分也只是提交一下便于回家修改

+++ GrapX/Platform/Win32_D3D11/GXGraphicsImpl_d3d11.cpp
@@ -942,7 +942,7 @@ namespace GrapX
         }
       }
 
-      FIMEMORY* fi_mem = FreeImage_OpenMemory((BYTE*)pBuffer->GetPtr(), pBuffer->GetSize());
+      FIMEMORY* fi_mem = FreeImage_OpenMemory((BYTE*)pBuffer->GetPtr(), (DWORD)pBuffer->GetSize());
       FREE_IMAGE_FORMAT fi_fmt = FreeImage_GetFileTypeFromMemory(fi_mem);
       if(fi_fmt == FIF_UNKNOWN) {
         return GX_ERROR_HANDLE;
